Audit specialized agnostic components
Entire-Checkpoint: 4c130ed75533
