+++ crates/ars-components/tests/spec_conformance.rs
@@ -38,7 +38,7 @@ mod overlay;
 #[path = "spec_conformance/selection.rs"]
 mod selection;
 
-#[path = "spec_conformance/specialized.rs"]
+#[path = "spec_conformance/specialized/mod.rs"]
 mod specialized;
 
 #[path = "spec_conformance/utility/mod.rs"]
