+++ crates/ars-components/tests/spec_conformance/specialized/clipboard.rs
@@ -1,0 +1,16 @@
+use super::{assert_anatomy, specialized_core};
+
+#[test]
+fn clipboard_anatomy_matches_spec() {
+    assert_anatomy(
+        "clipboard",
+        &[
+            (specialized_core::clipboard::Part::Root, "root"),
+            (specialized_core::clipboard::Part::Label, "label"),
+            (specialized_core::clipboard::Part::Trigger, "trigger"),
+            (specialized_core::clipboard::Part::Indicator, "indicator"),
+            (specialized_core::clipboard::Part::Status, "status"),
+            (specialized_core::clipboard::Part::ValueText, "value-text"),
+        ],
+    );
+}
